exclusão das pastas model e dao, DentistaService.java, PacienteService.java, ajustes na classes e interfaces relacionadas a dentista para funcionamento com ORM e configuração da application.properties

+++ src/main/java/com/example/sistemaDeReservasDeConsultas/repository/IDentistaRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.sistemaDeReservasDeConsultas.entity.DentistaEntity;
 
-public interface IDentistaRepository extends JpaRepository<DentistaEntity, Long> { }
+public interface IDentistaRepository extends JpaRepository<DentistaEntity, Integer> { }
